requirements.txt
beautifulsoup4
EbookLib
flet
fpdf2
googletrans
icrawler
pandas
pillow
pyinstaller
selenium
translators
undetected-chromedriver
urllib3
webdriver_manager
Proxy-List-Scrapper
undetected-edgedriver
zenrows
cloudscraper
django
djangorestframework
django-filter
markdown
pymongo[snappy,gssapi,srv,tls] == 3.12.3
dnspython
mongoengine
djongo